+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [unreleased]
 #### Added
+- Add `Crop & Export & Remove selected Images` action that allows to start a batch process that crops and exports the selected image(s) into a subfolder and also removes the image(s) from the list once it completed.
 - Selected images can now be removed from the list view via the context menu or using the backspace keyboard shortcut.
   - Note: The images are just removed from the list view in the UI, the images on the filesystem are untouched.